+++ app/schema/order.py
@@ -1,11 +1,13 @@
 from datetime import datetime
+from decimal import Decimal
+
 from pydantic import BaseModel
 
 
 class OrderItemResponse(BaseModel):
     product_id: int
     quantity: int
-    price_at_purchase: float
+    price_at_purchase: Decimal
 
     class Config:
         from_attributes = True
@@ -13,7 +15,7 @@ class OrderItemResponse(BaseModel):
 
 class OrderResponse(BaseModel):
     id: int
-    total_amount: float
+    total_amount: Decimal
     status: str
     created_at: datetime
     order_items: list[OrderItemResponse]
